Move loro document into own crdt module, refactor helper methods
